decompile rtc

+++ include/rtc.h
@@ -22,7 +22,7 @@ extern struct Time gLocalTime;
 void RtcDisableInterrupts(void);
 void RtcRestoreInterrupts(void);
 u32 ConvertBcdToBinary(u8 bcd);
-bool8 IsLeapYear(u8 year);
+bool8 IsLeapYear(u32 year);
 u16 ConvertDateToDayCount(u8 year, u8 month, u8 day);
 u16 RtcGetDayCount(struct SiiRtcInfo *rtc);
 void RtcInit(void);
